Pull ContractAnalysis + FunctionType from checker::type_check pass
