+++ src/vm/analysis/check_readonly_definitions/mod.rs
