+++ src/vm/analysis/check_typing/natives/options.rs
@@ -1,7 +1,7 @@
 use vm::representations::{SymbolicExpression};
 use vm::types::{AtomTypeIdentifier, TypeSignature};
 
-use vm::checker::typecheck::{TypeResult, TypingContext, 
+use vm::analysis::check_typing::{TypeResult, TypingContext, 
                              CheckError, CheckErrors, no_type, TypeChecker};
