+++ src/vm/analysis/diagnostic.rs
@@ -1,6 +1,6 @@
 use vm::representations::{Span};
 use std::fmt;
-use vm::checker::{CheckErrors};
+use vm::analysis::{CheckErrors};
 
 /// In a near future, we can go further in our static analysis and provide different levels 
 /// of diagnostics, such as warnings, hints, best practices, etc.
